requirements.txt
asgiref==3.5.2
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cycler==0.11.0
Django==4.0.5
django-jazzmin==2.5.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
Fiona==1.8.21
fonttools==4.33.3
geographiclib==1.52
geojson==2.5.0
geopandas==0.10.2
geopy==2.2.0
idna==3.3
kiwisolver==1.4.3
matplotlib==3.5.2
munch==2.5.0
networkx==2.8.4
numpy==1.22.4
osmnx==1.2.1
overpass==0.7
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
PyJWT==2.4.0
pyparsing==3.0.9
pyproj==3.3.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
Rtree==1.0.0
Shapely==1.8.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
